Support for vApp instantiation and removal
We begin with an implementation of a top-level method that instantiates
a new vApp from a template given by its ID in the target VDC. Custom
options may be provided by the calling method.

We also add support for retrieving individual catalog items. With this
new functionality it is possible to get catalog item from any catalog
and not just those that are present in the chosen one.

In order to support other operations on the vApp we also extend the
parsing of a vApp.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>