Remove basic network options during instantiation
Previous version included a network configuration in the
`InstantiationParams`, however it only allowed for specifying the
network name along with its href in a bridge fence mode. This patch
removes this and uses the network specification as given in the template
itself.

Specialisation is still possible by providing full `InstantiationParams`
object, including the network.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>