Support passing custom parser
With this commit we allow user to pass custom SAX parser
implementation to the `request()` call. Furthermore, we
support two special parsers:

- `plaintext` - return plaintext XML response. Until now
  ToHashDocument parser was always used which prevented user
  from getting the actual XML string.
- `xml` - return DOM-parsed XML response. This is an upgrade of
  `plaintext` option to make it more user-friendly.

Reason for such change is that RECONFIGURE requests rely on whole
XML sections being sent in body, even if we're not modifying them.
Therefore we need to be able to first obtain current XML description
of the VM to then only modify desired subsections.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>