Support virtual hardware reconfiguration for VM
With this commit we enhance the `post_reconfigure_vm` function to
actually support virtual hardware reconfiguration. vCloud requires
**ALL** the hardware configuration to be sent even if we're only modifying
a small piece of it. Current implementation of this call didn't respect
this limitation which resulted in hardware disappearing. When we updated
memory size, for example, all the VM disks vanished!

Syntax to configure hardware is now the very same to that of vapp template
instantiation:

```
options = {
    :name        => 'DB VM',
    :description => 'Some Description',
    :hardware    => {
        :memory => { :quantity_mb => 4096 },
        :cpu    => { :num_cores => 4, :cores_per_socket => 1 },
        :disk   => [
            { :id => '2000', :capacity_mb => 5*1024 },
            { :id => '2001', :capacity_mb => -1 },
            { :capacity_mb => 1*1024 }
        ]
    }
}
service.post_reconfigure_vm('vm-id', xml, options)
```

Documentation attached.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>