Implement vApp SAX parser
Until now default Hash parser was used when parsing vApp. With this
commit we rather implement custom SAX parser in order to be able to
reduce number of API requests made when iterating the cloud.

Basic observation we build on is that XML response of `service.get_vapp`
contains complete data also for child VMs. However, the way fog worked
until now was that it performed additional API call for each VM and only
parse it then...

With this commit we therefore implement (a very simple) custom SAX parser
for vApp and tune it to make use of existing Vm parser when parsing child
Vm XML. Some changes in models were required to support for such behavior
or else fog would perform API request for each VM.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>