Make sure vApp reloads when necessery
When vApp is parsed from the VDC summary XML it contains no
other data than vApp id and name. So we mustn't initialize vms
to empty array in order to trigger API request when this attribute
is accessed. Similar goes with vApp description: we now only set
it to empty string when we're certain that we're parsing the full-blown
vApp XML.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>