Prevent infinite loop when parsing VM without disks/NICs
If VM had no disks value NonLoaded was stored as a placeholder for
them which resulted in infinite loop whenever user wanted to access
this field. With this commit we prevent such disaster by defaulting
vm disks and NICs to empty array `[]` in case they did not get parsed
during the process (which implies they are not present).

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>